requirements.txt
flask
joblib
pdfminer.six
openai
flask_cors
pymongo
bcrypt
tinydb
flask_sqlalchemy